moved config parse sensor to plugin
