+++ dags/arctica1_file-drop.py
@@ -6,16 +6,13 @@ This DAG listens for files that should be dropped onto a mountpoint that is acce
 from airflow import utils
 from airflow import DAG
 
-from airflow.models import BaseOperator
 from airflow.operators.dummy_operator import DummyOperator
 from airflow.operators.python_operator import PythonOperator, ShortCircuitOperator
 from airflow.operators.bash_operator import BashOperator
 from airflow.operators.dagrun_operator import TriggerDagRunOperator
-from airflow.operators import EnsureDirectoryExistsOperator, RsyncOperator
+from airflow.operators import EnsureDirectoryExistsOperator, RsyncOperator, EnsureConfigurationExistsSensor
 
 from airflow.exceptions import AirflowException
-
-import yaml
 
 import logging
 LOG = logging.getLogger(__name__)
@@ -35,32 +32,9 @@ args = {
     'remove_files_after': 540, # minutes
 }
 
-
 ###
-# grab some experimental details that we need ###
+# define the workflow
 ###
-def parseConfigurationFile(**kwargs):
-    with open( kwargs['configuration_file'], 'r' ) as stream:
-        try:
-            d = yaml.load(stream)
-            for k,v in d.items():                
-                kwargs['task_instance'].xcom_push(key=k,value=v)
-            exp_dir = '%s_%s' % (d['experiment']['name'], d['experiment']['microscope'])
-            kwargs['task_instance'].xcom_push(key='experiment_directory', value='%s/%s/' % ( kwargs['destination_directory'], exp_dir))
-        except Exception as e:
-            raise AirflowException('Error creating destination directory: %s' % (e,))
-    return kwargs['configuration_file']
-class EnsureConfigurationExistsSensor(ShortCircuitOperator):
-    """ monitor for configuration defined on tem """
-    def __init__(self,configuration_file,destination_directory,*args,**kwargs):
-        super(EnsureConfigurationExistsSensor,self).__init__(
-            python_callable=parseConfigurationFile, 
-            op_kwargs={ 'configuration_file': configuration_file, 'destination_directory': destination_directory }, 
-            *args, **kwargs)
-            
-
-
-
 with DAG( 'cryoem_arctica1_file-drop',
         description="Monitor for new metadata and data on the arctica and put it into long term storage",
         schedule_interval="* * * * *",
