+++ plugins/cryoem_operators.py
@@ -1,0 +1,41 @@
+
+from airflow.plugins_manager import AirflowPlugin
+from airflow.utils.decorators import apply_defaults
+
+from airflow.operators.dummy_operator import DummyOperator
+from airflow.operators.python_operator import PythonOperator, ShortCircuitOperator
+
+
+import yaml
+
+
+import logging
+LOG = logging.getLogger(__name__)
+
+
+###
+# grab some experimental details that we need ###
+###
+def parseConfigurationFile(**kwargs):
+    with open( kwargs['configuration_file'], 'r' ) as stream:
+        try:
+            d = yaml.load(stream)
+            for k,v in d.items():                
+                kwargs['task_instance'].xcom_push(key=k,value=v)
+            exp_dir = '%s_%s' % (d['experiment']['name'], d['experiment']['microscope'])
+            kwargs['task_instance'].xcom_push(key='experiment_directory', value='%s/%s/' % ( kwargs['destination_directory'], exp_dir))
+        except Exception as e:
+            raise AirflowException('Error creating destination directory: %s' % (e,))
+    return kwargs['configuration_file']
+class EnsureConfigurationExistsSensor(ShortCircuitOperator):
+    """ monitor for configuration defined on tem """
+    def __init__(self,configuration_file,destination_directory,*args,**kwargs):
+        super(EnsureConfigurationExistsSensor,self).__init__(
+            python_callable=parseConfigurationFile, 
+            op_kwargs={ 'configuration_file': configuration_file, 'destination_directory': destination_directory }, 
+            *args, **kwargs)
+
+
+class FilePlugin(AirflowPlugin):
+    name = 'cryoem_plugin'
+    operators = [EnsureConfigurationExistsSensor,]
